@@ -1,10 +1,10 @@
 name: daemon
-description: Enjin Wallet Daemon for Windows, macOS, and Linux
+description: Enjin Wallet Daemon is a cross-platform desktop application that allows game developers to interact with the Enjin Platform
 publish_to: 'none'
-version: 2.1.1+1
+version: 2.3.0
 
 environment:
-  sdk: ^3.4.0
+  sdk: ^3.6.0
 
 dependencies:
   flutter:
@@ -12,25 +12,26 @@ dependencies:
   uuid: ^4.5.1
   path: ^1.9.1
   path_provider: ^2.1.5
-  download_assets: ^3.2.0
+  background_downloader: ^8.9.4
+  archive: ^4.0.2
   flutter_svg: ^2.0.17
   xterm: ^4.0.0
-  dio: ^5.7.0
+  dio: ^5.8.0+1
   convert: ^3.1.2
-  lottie: ^3.3.0
-  sembast: ^3.8.1+1
+  lottie: ^3.3.1
+  sembast: ^3.8.3
   encrypt: ^5.0.3
   crypto: ^3.0.6
   window_manager: ^0.4.3
   get: ^4.6.6
   outline_gradient_button: ^2.0.1+1
-  intl: ^0.20.1
+  intl: ^0.20.2
   aligned_dialog: ^0.0.7
   flutter_animate: ^4.5.2
 
 dev_dependencies:
-  flutter_lints: ^4.0.0
-  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^5.0.0
+  flutter_launcher_icons: ^0.14.3
 
 flutter_launcher_icons:
   image_path: "lib/assets/enjin_logo.png"
