@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class BaseButton extends StatelessWidget {
   const BaseButton({
-    Key? key,
+    super.key,
     required this.text,
     this.onPressed,
     this.buttonStyle,
@@ -12,9 +12,7 @@ class BaseButton extends StatelessWidget {
     this.width,
     this.margin,
     this.alignment,
-  }) : super(
-          key: key,
-        );
+  });
 
   final String text;
 
