@@ -8,26 +8,16 @@ class CustomElevatedButton extends BaseButton {
     this.decoration,
     this.leftIcon,
     this.rightIcon,
-    EdgeInsets? margin,
-    VoidCallback? onPressed,
-    ButtonStyle? buttonStyle,
-    Alignment? alignment,
-    TextStyle? buttonTextStyle,
-    bool? isDisabled,
-    double? height,
-    double? width,
-    required String text,
-  }) : super(
-          text: text,
-          onPressed: onPressed,
-          buttonStyle: buttonStyle,
-          isDisabled: isDisabled,
-          buttonTextStyle: buttonTextStyle,
-          height: height,
-          width: width,
-          alignment: alignment,
-          margin: margin,
-        );
+    super.margin,
+    super.onPressed,
+    super.buttonStyle,
+    super.alignment,
+    super.buttonTextStyle,
+    super.isDisabled,
+    super.height,
+    super.width,
+    required super.text,
+  });
 
   final BoxDecoration? decoration;
 
